Move Symptoms into Stories
