+++ backend/alembic/env.py
@@ -7,7 +7,6 @@ from sqlalchemy import engine_from_config, pool
 from database.database import Base
 from users import models as user_models
 from stories import models as story_models
-from symptoms import models as symptoms_models
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
