+++ backend/stories/models.py
@@ -21,7 +21,7 @@ class Story(Base):
     sickness_end = Column(String(64))
     current_location = Column(String(128))
     user = relationship("User", uselist=False, back_populates="story")
-    symptoms = relationship("Symptom", secondary="story_symptoms", back_populates="story")
+    symptoms = relationship("Symptom", secondary="story_symptoms")
 
     @property
     def medical_problems(self):
@@ -30,3 +30,15 @@ class Story(Base):
     @medical_problems.setter
     def medical_problems(self, value):
         self._medical_problems = json.dumps(value)
+
+story_symptoms = Table('story_symptoms', Base.metadata,
+    Column('id', Integer, primary_key=True, index=True),
+    Column('story_id', ForeignKey('stories.id')),
+    Column('symptom_id', ForeignKey('symptoms.id'))
+)
+
+class Symptom(Base):
+    __tablename__ = "symptoms"
+
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String(128), unique=True, index=True)
